installer layout for susi
Separate the installer from the susi_linux application, and
factor stuff that is RPi specific into the `raspi` sub-directory.

Change the layout of the installed files on RPi:
/home/pi/SUSI.AI/
	config.json
	susi_installer/
	susi_linux/
	susi_server/
	susi_skill_data/
	susi_api_wrapper/
	seeed_voicecard/
The CWD of all programs are changed to /home/pi/SUSI.AI so that
the configuration file there is used.

Change to systemd service files to the new layout.

Adjust the factory_reset.sh script to the new layout.

The ss-susi-linux@pi service is automatically activated on startup,
so do not enable/disable the ss-susi-linux.service in the access_point
server and (r)wap.sh scripts.