introduce central configuration program susi-config
This PR introduces a program susi-config which is responsible for
setting all kind of parameters in the config.json file, as well as logging
into the susi.ai services.

It replaces susi_linux scripts authentication, config_generator, and susi_login.

The flask AP server is adjusted to use the new script.

All calls are done as user pi instead of root to make sure that config.json
always remains owned by pi.